Refactor out `lovatAPI.getNotesByTeamNumber()`
