+++ lib/analysis_functions/team_lookup_notes_analysis.dart
@@ -1,5 +1,6 @@
 import 'package:scouting_dashboard_app/analysis_functions/analysis.dart';
 import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/team_lookup/get_notes.dart';
 
 class TeamLookupNotesAnalysis extends AnalysisFunction {
   TeamLookupNotesAnalysis({
